[internals] move loadStyleSheets to internals and use it in data grid and charts
