+++ packages/x-internals/src/export/index.ts
@@ -1,0 +1,1 @@
+export { loadStyleSheets } from './loadStyleSheets';
